+++ src/AdminConsole/Components/Pages/App/Settings/Authenticators.razor
@@ -1,6 +1,7 @@
 @page "/app/{app}/settings/authenticators"
 @using Passwordless.Common.Models.MDS
 @using Passwordless.Common.Models.Apps
+@using Passwordless.Common.Models.Authenticators
 
 @inject NavigationManager NavigationManager
 @inject ICurrentContext CurrentContext
