+++ src/AdminConsole/Components/Pages/App/Settings/ManageAuthenticators.razor
@@ -1,5 +1,6 @@
 @page "/app/{app}/settings/authenticators/manage"
 @using Passwordless.Common.Models.Apps
+@using Passwordless.Common.Models.Authenticators
 @using Passwordless.Common.Models.MDS
 
 @inherits BaseApplicationPage
